Refactor ETL pipeline and address review feedback
- Move DataProcessor to etl/processor.py for better separation of concerns.
- Robustify WSDC ID extraction with path-based fallback.
- Improve error handling and schema validation in query_ledger.py.
- Fix couples data parsing in ScoringDanceParser.
- Ensure content/studies/ is clean of test artifacts.
- Update tests to reflect architectural changes.
