+++ etl/processor.py
@@ -1,0 +1,16 @@
+import pandas as pd
+
+class DataProcessor:
+    """Handles data transformation and aggregation."""
+    @staticmethod
+    def process_for_ledger(raw_df):
+        if raw_df.empty:
+            return pd.DataFrame()
+
+        # Group by the new Dancer_ID and other metadata
+        processed_df = raw_df.groupby(['Dancer_ID', 'competitor_name', 'result_id', 'event_title', 'event_date']).agg(
+            Registry_Points_Sum=('wsdc_points', 'sum'),
+            Promoted=('Promoted', 'any')
+        ).reset_index()
+
+        return processed_df
